printing

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -7,7 +7,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	size_t node = 0;
-	long int list;
+	int list;
 
 	while (head)
 	{
